moved the LagLeadDataView inside the main LagLeadOperator class.
